[refs #00184] Move font-size: 0; hack
The font-size hack that is used to remove whitespace in the nav was
leaking out into other (namely button) styles. I’ve moved the hack to
sit directly on the nav’s list rather than its wrapping element. This
means that we don’t need to undo its affects on any other elements are
we did previously.