[refs #00184] Swap search button for proper sprite
We’ve reworked the main search button to use an actual BUTTON element,
as well as using the spriting element for the icon. Also some nice
sibling selector trickery to do away with the -ve margin that made the
button sit above the search field: they’re now back next to one another
and styled a little more cleanly.