[refs #00271] Remove Safari Repaint Hack
Safari exhibited an unusual bug whereby any rem-sized text appearing
anywhere after a DETAILS element was rendered impossibly small. Like,
1px kinda size. The/a way of getting around it was to force a full-page
repaint, which is icky, but it was scoped to only Safari users.

Anyway! Good news is that the bug has been fixed[1] so we can delete the
hack, which is nice.

1. https://bugs.webkit.org/show_bug.cgi?id=173876